requirements.txt
fastapi==0.89.1
uvicorn==0.20.0
python-dotenv==0.21.1
sqlalchemy
sqlmodel
pyjwt